ファイルの削除: Hero3DViewおよびHeroSideBarコンポーネントを削除し、レイアウトを更新
