+++ app/(Home)/Hero3DView.tsx
