+++ app/(Home)/HeroSideBar.tsx
